requirements/base.txt
# Shared across all services
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
requests==2.32.4
openai==1.99.1